appendNextStringValue() gets a buffer.
This is to deal with situations where long string data needs to be written
to a file or piped through some other Writer that performs synchronization.
The idea is to reduce the time we go through a synchronization lock without
having to constantly allocate and throw away a buffer. The buffer lives for
the duration of the stream parse.