Added parseUnicodeEscape() to lexer
This allows incremental parsing of String \u escape sequences. If any of the following 4 characters are not hexadecimal, abort immediately. This gives a more accurate location for a parse exception, and doesn't advance the reader beyond the error site.

Easier for the lexer, stream reader, etc to deal with RFC 7464 -- JavaScript Object Notation (JSON) Text Sequences if a text sequence is truncated.