Add Collectors for structure creation.
Add StructureCollector, with implementation classes JSONCollector and
CollectionCollector to accumule JSON result objects for both org.json and
java.util container objects. The Collector is similar to Java 8 streams.

The trampoline code is adjusted to be able to return object results at the
end of a structure parse, rather than eager addition at the beginning.